Performance related: get endpoint using new way
Using appMgr.kubeClient is slow. This becomes very
obvious when the number of resources gets big.
Because GetEndpoints(getEndpoints) is called many times when a new
request is coming during processCfgMap.

When the number of eps is relatively small, for example, in one of my
environment, when the number is in around 30,
the old way using kubeClient to GetEndpoints() is about 0.2 seconds,
while it is 0.000005s using the new way.

Example:
old way: from 02-23T15:10:10.463027031Z to 02-23T15:10:10.661902573Z;
new way: from 02-23T15:10:10.661943586Z to 02-23T15:10:10.66194896Z.

Because this is called many times so it is gonna be 0.2*n seconds
vs 0.000005*n, the difference is huge when n becomes 100, 1000, etc.